requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.21.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Data validation and modeling
pydantic>=2.0.0

# Enhanced features (optional)
# requests>=2.28.0  # For WhatsApp Business API
# sqlalchemy>=2.0.0  # For database storage
# redis>=4.0.0      # For caching
# celery>=5.2.0     # For background tasks

